add additional helper functions for the queue for data retrieval
