+++ pkg/execution/state/redis_state/queue.go
@@ -9,7 +9,6 @@ import (
 	"iter"
 	"math"
 	mrand "math/rand"
-	"math/rand/v2"
 	"strconv"
 	"strings"
 	"sync"
@@ -164,6 +163,7 @@ type QueueManager interface {
 	osqueue.Queue
 	osqueue.QueueDirectAccess
 
+	DequeueByJobID(ctx context.Context, jobID string, opts ...QueueOpOpt) error
 	Dequeue(ctx context.Context, queueShard QueueShard, i osqueue.QueueItem) error
 	Requeue(ctx context.Context, queueShard QueueShard, i osqueue.QueueItem, at time.Time) error
 	RequeueByJobID(ctx context.Context, queueShard QueueShard, jobID string, at time.Time) error
@@ -178,8 +178,10 @@ type QueueManager interface {
 	BacklogsByPartition(ctx context.Context, queueShard QueueShard, partitionID string, from time.Time, until time.Time, opts ...QueueIterOpt) (iter.Seq[*QueueBacklog], error)
 	// BacklogSize retrieves the number of items in the specified backlog
 	BacklogSize(ctx context.Context, queueShard QueueShard, backlogID string) (int64, error)
-
+	// PartitionByID retrieves the partition by the partition ID
 	PartitionByID(ctx context.Context, queueShard QueueShard, partitionID string) (*PartitionInspectionResult, error)
+	// ItemByID retrieves the queue item by the jobID
+	ItemByID(ctx context.Context, jobID string, opts ...QueueOpOpt) (*osqueue.QueueItem, error)
 
 	// PartitionBacklogSize returns the point in time backlog size of the partition.
 	// This will sum the size of all backlogs in that partition
@@ -1492,7 +1494,6 @@ func (q *queue) dropPartitionPointerIfEmpty(ctx context.Context, shard QueueShar
 	}
 }
 
-
 func (q *queue) SetFunctionMigrate(ctx context.Context, sourceShard string, fnID uuid.UUID, migrateLockUntil *time.Time) error {
 	ctx = redis_telemetry.WithScope(redis_telemetry.WithOpName(ctx, "SetFunctionMigrate"), redis_telemetry.ScopeQueue)
 
@@ -2281,297 +2282,6 @@ func (q *queue) ExtendLease(ctx context.Context, i osqueue.QueueItem, leaseID ul
 		return nil, ErrQueueItemLeaseMismatch
 	default:
 		return nil, fmt.Errorf("unknown response extending lease: %d", status)
-	}
-}
-
-// Dequeue removes an item from the queue entirely.
-func (q *queue) Dequeue(ctx context.Context, queueShard QueueShard, i osqueue.QueueItem) error {
-	ctx = redis_telemetry.WithScope(redis_telemetry.WithOpName(ctx, "Dequeue"), redis_telemetry.ScopeQueue)
-
-	if queueShard.Kind != string(enums.QueueShardKindRedis) {
-		return fmt.Errorf("unsupported queue shard kind for Dequeue: %s", queueShard.Kind)
-	}
-
-	kg := queueShard.RedisClient.kg
-
-	partition := q.ItemShadowPartition(ctx, i)
-	backlog := q.ItemBacklog(ctx, i)
-
-	keys := []string{
-		kg.QueueItem(),
-		kg.PartitionItem(),
-
-		kg.ConcurrencyIndex(),
-
-		partition.readyQueueKey(kg),
-		kg.GlobalPartitionIndex(),
-		kg.GlobalAccountIndex(),
-		kg.AccountPartitionIndex(i.Data.Identifier.AccountID),
-
-		kg.ShadowPartitionMeta(),
-		kg.BacklogMeta(),
-
-		kg.BacklogSet(backlog.BacklogID),
-		kg.ShadowPartitionSet(partition.PartitionID),
-		kg.GlobalShadowPartitionSet(),
-		kg.GlobalAccountShadowPartitions(),
-		kg.AccountShadowPartitions(i.Data.Identifier.AccountID),
-		kg.PartitionNormalizeSet(partition.PartitionID),
-
-		// In progress keys
-		partition.accountInProgressKey(kg),
-		partition.inProgressKey(kg),
-		backlog.customKeyInProgress(kg, 1),
-		backlog.customKeyInProgress(kg, 2),
-
-		// Active set keys
-		partition.accountActiveKey(kg),
-		partition.activeKey(kg),
-		backlog.customKeyActive(kg, 1),
-		backlog.customKeyActive(kg, 2),
-		backlog.activeKey(kg),
-
-		// Active run sets
-		kg.RunActiveSet(i.Data.Identifier.RunID), // Set for active items in run
-		partition.accountActiveRunKey(kg),        // Set for active runs in account
-		partition.activeRunKey(kg),               // Set for active runs in partition
-		backlog.customKeyActiveRuns(kg, 1),       // Set for active runs with custom concurrency key 1
-		backlog.customKeyActiveRuns(kg, 2),       // Set for active runs with custom concurrency key 2
-
-		kg.Idempotency(i.ID),
-
-		// Singleton
-		kg.SingletonRunKey(i.Data.Identifier.RunID.String()),
-	}
-
-	// Append indexes
-	for _, idx := range q.itemIndexer(ctx, i, queueShard.RedisClient.kg) {
-		if idx != "" {
-			keys = append(keys, idx)
-		}
-	}
-
-	idempotency := q.idempotencyTTL
-	if q.idempotencyTTLFunc != nil {
-		idempotency = q.idempotencyTTLFunc(ctx, i)
-	}
-
-	args, err := StrSlice([]any{
-		i.ID,
-		partition.PartitionID,
-		backlog.BacklogID,
-		i.Data.Identifier.AccountID.String(),
-		i.Data.Identifier.RunID.String(),
-
-		int(idempotency.Seconds()),
-	})
-	if err != nil {
-		return err
-	}
-
-	status, err := scripts["queue/dequeue"].Exec(
-		redis_telemetry.WithScriptName(ctx, "dequeue"),
-		queueShard.RedisClient.unshardedRc,
-		keys,
-		args,
-	).AsInt64()
-	if err != nil {
-		return fmt.Errorf("error dequeueing item: %w", err)
-	}
-	switch status {
-	case 0:
-		if rand.Float64() < 0.05 {
-			q.log.Debug("dequeued item", "job_id", i.ID, "item", i)
-		}
-
-		return nil
-	case 1:
-		return ErrQueueItemNotFound
-	default:
-		return fmt.Errorf("unknown response dequeueing item: %d", status)
-	}
-}
-
-// Requeue requeues an item in the future.
-func (q *queue) Requeue(ctx context.Context, queueShard QueueShard, i osqueue.QueueItem, at time.Time) error {
-	ctx = redis_telemetry.WithScope(redis_telemetry.WithOpName(ctx, "Requeue"), redis_telemetry.ScopeQueue)
-
-	l := q.log.With("item", i)
-
-	if queueShard.Kind != string(enums.QueueShardKindRedis) {
-		return fmt.Errorf("unsupported queue shard kind for Requeue: %s", queueShard.Kind)
-	}
-
-	kg := queueShard.RedisClient.kg
-
-	now := q.clock.Now()
-	if at.Before(now) {
-		at = now
-	}
-
-	// Unset any lease ID as this is requeued.
-	i.LeaseID = nil
-	// Update the At timestamp.
-	// NOTE: This does no priority factorization or FIFO for function ordering,
-	// eg. adjusting AtMS based off of function run time.
-	i.AtMS = at.UnixMilli()
-	// Update the wall time that this should run at.
-	i.WallTimeMS = at.UnixMilli()
-
-	// Reset refill details
-	i.RefilledFrom = ""
-	i.RefilledAt = 0
-
-	// Reset enqueuedAt (used for latency calculation)
-	i.EnqueuedAt = now.UnixMilli()
-
-	fnPartition := q.ItemPartition(ctx, queueShard, i)
-	shadowPartition := q.ItemShadowPartition(ctx, i)
-
-	requeueToBacklog := q.itemEnableKeyQueues(ctx, i)
-
-	requeueToBacklogsVal := "0"
-	if requeueToBacklog {
-		requeueToBacklogsVal = "1"
-
-		// To avoid requeueing item into a stale backlog, retrieve latest throttle
-		if i.Data.Throttle != nil && i.Data.Throttle.KeyExpressionHash == "" {
-			refreshedThrottle, err := q.refreshItemThrottle(ctx, &i)
-			if err != nil {
-				// If we cannot find the event for the queue item, dequeue it. The state
-				// must exist for the entire duration of a function run.
-				if errors.Is(err, state.ErrEventNotFound) {
-					l.Warn("could not find event for refreshing throttle before requeue")
-
-					err := q.Dequeue(ctx, queueShard, i)
-					if err != nil && !errors.Is(err, ErrQueueItemNotFound) {
-						return fmt.Errorf("could not dequeue item with missing throttle state: %w", err)
-					}
-
-					return nil
-				}
-
-				return fmt.Errorf("could not refresh item throttle: %w", err)
-			}
-
-			// Update throttle to latest evaluated value + expression hash
-			i.Data.Throttle = refreshedThrottle
-		}
-	}
-
-	backlog := q.ItemBacklog(ctx, i)
-
-	keys := []string{
-		kg.QueueItem(),
-		kg.PartitionItem(), // Partition item, map
-		kg.ConcurrencyIndex(),
-
-		kg.GlobalPartitionIndex(),
-		kg.GlobalAccountIndex(),
-		kg.AccountPartitionIndex(i.Data.Identifier.AccountID),
-
-		shadowPartition.readyQueueKey(kg),
-
-		// In progress (concurrency) keys
-		shadowPartition.accountInProgressKey(kg),
-		shadowPartition.inProgressKey(kg),
-		backlog.customKeyInProgress(kg, 1),
-		backlog.customKeyInProgress(kg, 2),
-
-		// Active set keys
-		shadowPartition.accountActiveKey(kg),
-		shadowPartition.activeKey(kg),
-		backlog.customKeyActive(kg, 1),
-		backlog.customKeyActive(kg, 2),
-		backlog.activeKey(kg),
-
-		// Active run sets
-		kg.RunActiveSet(i.Data.Identifier.RunID), // Set for active items in run
-		shadowPartition.accountActiveRunKey(kg),  // Set for active runs in account
-		shadowPartition.activeRunKey(kg),         // Set for active runs in partition
-		backlog.customKeyActiveRuns(kg, 1),       // Set for active runs with custom concurrency key 1
-		backlog.customKeyActiveRuns(kg, 2),       // Set for active runs with custom concurrency key 2
-
-		// key queues v2
-		kg.BacklogSet(backlog.BacklogID),
-		kg.BacklogMeta(),
-		kg.GlobalShadowPartitionSet(),
-		kg.ShadowPartitionSet(shadowPartition.PartitionID),
-		kg.ShadowPartitionMeta(),
-		kg.GlobalAccountShadowPartitions(),
-		kg.AccountShadowPartitions(i.Data.Identifier.AccountID), // empty for system partitions
-	}
-	// Append indexes
-	for _, idx := range q.itemIndexer(ctx, i, queueShard.RedisClient.kg) {
-		if idx != "" {
-			keys = append(keys, idx)
-		}
-	}
-
-	args, err := StrSlice([]any{
-		i.ID,
-		i,
-		at.UnixMilli(),
-
-		i.Data.Identifier.AccountID.String(),
-		i.Data.Identifier.RunID.String(),
-		fnPartition.ID,
-		fnPartition,
-
-		now.UnixMilli(),
-
-		requeueToBacklogsVal,
-		shadowPartition,
-		backlog.BacklogID,
-		backlog,
-	})
-	if err != nil {
-		return err
-	}
-
-	q.log.Trace("requeueing queue item",
-		"id", i.ID,
-		"kind", i.Data.Kind,
-		"time", at.Format(time.StampMilli),
-		"partition_id", shadowPartition.PartitionID,
-		"backlog", requeueToBacklogsVal,
-	)
-
-	status, err := scripts["queue/requeue"].Exec(
-		redis_telemetry.WithScriptName(ctx, "requeue"),
-		queueShard.RedisClient.unshardedRc,
-		keys,
-		args,
-	).AsInt64()
-	if err != nil {
-		q.log.Error("error requeueing queue item",
-			"error", err,
-			"item", i,
-			"partition", fnPartition,
-			"shadow", shadowPartition,
-		)
-		return fmt.Errorf("error requeueing item: %w", err)
-	}
-	switch status {
-	case 0:
-		switch requeueToBacklogsVal {
-		case "1":
-			metrics.IncrBacklogRequeuedCounter(ctx, metrics.CounterOpt{
-				PkgName: pkgName,
-				Tags: map[string]any{
-					"queue_shard": q.primaryQueueShard.Name,
-					// "partition_id": i.FunctionID.String(),
-				},
-			})
-		}
-
-		return nil
-	case 1:
-		// This should only ever happen if a run is cancelled and all queue items
-		// are deleted before requeueing.
-		return ErrQueueItemNotFound
-	default:
-		return fmt.Errorf("unknown response requeueing item: %v (%T)", status, status)
 	}
 }
