+++ pkg/execution/state/redis_state/reader.go
@@ -590,3 +590,33 @@ func (q *queue) QueueIterator(ctx context.Context, opts QueueIteratorOpts) (part
 
 	return atomic.LoadInt64(&totalPartitions), atomic.LoadInt64(&totalQueueItems), nil
 }
+
+func (q *queue) ItemByID(ctx context.Context, jobID string, opts ...QueueOpOpt) (*osqueue.QueueItem, error) {
+	opt := newQueueOpOpt()
+	for _, apply := range opts {
+		apply(&opt)
+	}
+
+	shard := q.primaryQueueShard
+	if opt.shard != nil {
+		shard = *opt.shard
+	}
+
+	rc := shard.RedisClient.Client()
+	kg := shard.RedisClient.kg
+
+	cmd := rc.B().Hget().Key(kg.QueueItem()).Field(jobID).Build()
+	byt, err := rc.Do(ctx, cmd).AsBytes()
+	if err != nil {
+		if rueidis.IsRedisNil(err) {
+			return nil, ErrQueueItemNotFound
+		}
+	}
+
+	var item osqueue.QueueItem
+	if err := json.Unmarshal(byt, &item); err != nil {
+		return nil, fmt.Errorf("error unmarshalling queue item: %w", err)
+	}
+
+	return &item, nil
+}
