+++ pkg/execution/state/redis_state/reader_test.go
@@ -6,12 +6,14 @@ import (
 	"testing"
 	"time"
 
+	"crypto/rand"
 	"github.com/google/uuid"
 	"github.com/inngest/inngest/pkg/consts"
 	"github.com/inngest/inngest/pkg/enums"
 	osqueue "github.com/inngest/inngest/pkg/execution/queue"
 	"github.com/inngest/inngest/pkg/execution/state"
 	"github.com/jonboulle/clockwork"
+	"github.com/oklog/ulid/v2"
 	"github.com/stretchr/testify/require"
 )
 
@@ -359,4 +361,67 @@ func TestQueueIterator(t *testing.T) {
 			require.EqualValues(t, tc.items, piCnt)
 		})
 	}
+}
+
+func TestItemByID(t *testing.T) {
+	_, rc := initRedis(t)
+	defer rc.Close()
+
+	ctx := context.Background()
+	clock := clockwork.NewFakeClock()
+	defaultShard := QueueShard{
+		Kind:        string(enums.QueueShardKindRedis),
+		RedisClient: NewQueueClient(rc, QueueDefaultKey),
+		Name:        consts.DefaultQueueShardName,
+	}
+	acctId, fnID, wsID := uuid.New(), uuid.New(), uuid.New()
+
+	q1 := NewQueue(
+		defaultShard,
+		WithAllowKeyQueues(func(ctx context.Context, acctID uuid.UUID) bool {
+			return false
+		}),
+		WithDisableLeaseChecks(func(ctx context.Context, acctID uuid.UUID) bool {
+			return false
+		}),
+		WithClock(clock),
+	)
+
+	enqueue := func(ctx context.Context, shard QueueShard) (osqueue.QueueItem, error) {
+		item := osqueue.QueueItem{
+			ID:          ulid.MustNew(ulid.Now(), rand.Reader).String(),
+			FunctionID:  fnID,
+			WorkspaceID: wsID,
+			Data: osqueue.Item{
+				WorkspaceID: wsID,
+				Kind:        osqueue.KindEdge,
+				Identifier: state.Identifier{
+					AccountID:   acctId,
+					WorkspaceID: wsID,
+					WorkflowID:  fnID,
+				},
+			},
+		}
+
+		return q1.EnqueueItem(ctx, shard, item, clock.Now(), osqueue.EnqueueOpts{})
+	}
+
+	t.Run("should be able to find the queue item", func(t *testing.T) {
+		enqueued, err := enqueue(ctx, defaultShard)
+		require.NoError(t, err)
+
+		res, err := q1.ItemByID(ctx, enqueued.ID)
+		require.NoError(t, err)
+		require.NotNil(t, res)
+		require.Equal(t, enqueued.ID, res.ID)
+	})
+
+	t.Run("should return not found error if absent", func(t *testing.T) {
+		_, err := enqueue(ctx, defaultShard)
+		require.NoError(t, err)
+
+		res, err := q1.ItemByID(ctx, "random")
+		require.ErrorIs(t, err, ErrQueueItemNotFound)
+		require.Nil(t, res)
+	})
 }
